use hive 2.3.4 so that we only need to copy 1 file
